Deletando arquivos desnecessarios

+++ src/main/java/school/sptech/Main.java
@@ -114,28 +114,30 @@ public class Main {
 
             for (S3Object object : objects) {
 
-                Path oupuPath = new File(object.key()).toPath();
+                if (object.key().endsWith(".xlsx")) {
+                    Path oupuPath = new File(object.key()).toPath();
 
-                if(Files.exists(oupuPath)){
-                    Files.delete(oupuPath);
+                    if(Files.exists(oupuPath)){
+                        Files.delete(oupuPath);
+                    }
+
+
+                    GetObjectRequest getObjectRequest = GetObjectRequest.builder()
+                            .bucket(nomeBucket)
+                            .key(object.key())
+                            .build();
+
+                    try (InputStream inputStream = s3Client.getObject(getObjectRequest, ResponseTransformer.toInputStream())) {
+                        Path outputPath = new File(object.key()).toPath();
+                        Files.copy(inputStream, outputPath);
+                        System.out.println("Arquivo baixado: " + object.key());
+                        Log.inserirNoLog("["+ LocalDateTime.now() .format(formatter)+ "] Arquivo baixado com sucesso: " + object.key());
+                    } catch (IOException e) {
+
+                        Log.inserirNoLog("["+ LocalDateTime.now() .format(formatter)+ "] Erro ao salvar o arquivo: " + e.getMessage());
+                        System.err.println("Erro ao salvar o arquivo: " + e.getMessage());
+                    }
                 }
-
-                GetObjectRequest getObjectRequest = GetObjectRequest.builder()
-                        .bucket(nomeBucket)
-                        .key(object.key())
-                        .build();
-
-                try (InputStream inputStream = s3Client.getObject(getObjectRequest, ResponseTransformer.toInputStream())) {
-                    Path outputPath = new File(object.key()).toPath();
-                    Files.copy(inputStream, outputPath);
-                    System.out.println("Arquivo baixado: " + object.key());
-                    Log.inserirNoLog("["+ LocalDateTime.now() .format(formatter)+ "] Arquivo baixado com sucesso: " + object.key());
-                } catch (IOException e) {
-
-                    Log.inserirNoLog("["+ LocalDateTime.now() .format(formatter)+ "] Erro ao salvar o arquivo: " + e.getMessage());
-                    System.err.println("Erro ao salvar o arquivo: " + e.getMessage());
-                }
-
             }
         } catch (S3Exception e) {
 
@@ -220,17 +222,16 @@ public class Main {
         System.out.println("Dados de população do Espirito Santos inseridos com sucesso");
         Log.inserirNoLog("["+ LocalDateTime.now() .format(formatter)+ "] Dados de população do Espirito Santos inseridos com sucesso");
 
-            //for (Dados dados : dadosEstraidos) {
-//                connection.update("INSERT INTO dados (dataFurto, horario, tipoObjeto, municipio) VALUES (?, ?, ?, ?)",
-//
-//                        dados.getData(),
-//                        dados.getHorario(),
-//                        dados.getObjeto(),
-//                        dados.getMunicipio());
+            for (Dados dados : dadosEstraidos) {
+                connection.update("INSERT INTO dados (dataFurto, horario, tipoObjeto, municipio) VALUES (?, ?, ?, ?)",
+                        dados.getData(),
+                        dados.getHorario(),
+                        dados.getObjeto(),
+                        dados.getMunicipio());
                 connection.update("INSERT INTO dados (dataFurto, horario, tipoObjeto, municipio) VALUES (?, ?, ?, ?)",
                 "2024-04-06", "00:00", "CELULAR", "VITORIA");
                 //System.out.println(dados);
-            //}
+            }
 
             System.out.println("Dados sobre furtos inseridos com sucesso no banco de dados!");
             Log.inserirNoLog("["+ LocalDateTime.now() .format(formatter)+ "] Dados sobre furtos inseridos com sucesso no banco de dados!");
